+++ app/Models/Brand.php
@@ -30,4 +30,12 @@ class Brand extends Model
     {
         return $this->hasMany(Branch::class);
     }
+
+    /**
+     * @return HasMany<Invitation, $this>
+     */
+    public function invitations(): HasMany
+    {
+        return $this->hasMany(Invitation::class);
+    }
 }