+++ app/Models/Organization.php
@@ -61,6 +61,14 @@ class Organization extends Model
     }
 
     /**
+     * @return HasMany<Invitation, $this>
+     */
+    public function invitations(): HasMany
+    {
+        return $this->hasMany(Invitation::class);
+    }
+
+    /**
      * @return BelongsToMany<User, $this>
      */
     public function activeUsers(): BelongsToMany